requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
geopy>=2.3.0
pytz>=2023.3
joblib>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
